Autotoolify and pkgconfigy.
In addition to cmake build files, this introduces autoconf/automake
files needed as well as pkgconfig files to the project. This also
completely cleanly finishes `make distcheck` and can be used to
create official release tarballs with `make dist` that should allow
all Linux distributions to `./configure && make && make install`
permitted the dependencies are met, but those are properly checked
by the configure script and nicely use pkgconfig for detection.

This should significantly aid in getting this package adopted
permitted that release tarballs are cut and posted to github
with every release.